requirements.txt
appdirs==1.4.4
black==24.2.0
certifi==2024.2.2
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configmypy==0.1.0
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
mypy-extensions==1.0.0
netCDF4==1.6.5
networkx==3.2.1
neuraloperator==0.3.0
nose==1.3.7
numpy==1.26.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.1
pytest==8.0.0
pytest-mock==3.12.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.12.0
sentry-sdk==1.40.4
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorly==0.8.1
tensorly-torch==0.4.0
torch==2.2.0
torch-harmonics==0.6.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
-e git+https://github.com/Luca-DTU/FNO-Wake-fidelity-Enhancement.git@b21b6ce496409c1181bba13ff96455c894373ddd#egg=WakeFNO
wandb==0.16.3
xarray==2024.2.0
